backend/requirements.txt
openai-whisper
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
aiofiles